refactor: rename service files and update import paths for consistency
