+++ src/services/index.ts
@@ -1,5 +1,5 @@
-export * from './Project';
-export * from './Tag';
-export * from './Task';
-export * from './Status';
+export * from './Project.service';
+export * from './Tag.service';
+export * from './Task.service';
+export * from './Status.service';
 export * from './User';
